src/main/java/org/example/MergeSort.java
package org.example;

public class MergeSort {

    // Method to sort an array using Merge Sort
    public static void sort(int[] array) {
        if (array == null || array.length < 2) {
            return;
        }
        mergeSort(array, new int[array.length], 0, array.length - 1);
    }

    // Recursive method for merge sort
    private static void mergeSort(int[] array, int[] temp, int leftStart, int rightEnd) {
        if (leftStart >= rightEnd) {
            return;
        }
        int middle = (leftStart + rightEnd) / 2;
        mergeSort(array, temp, leftStart, middle);
        mergeSort(array, temp, middle + 1, rightEnd);
        mergeHalves(array, temp, leftStart, rightEnd);
    }

    // Method to merge two halves
    private static void mergeHalves(int[] array, int[] temp, int leftStart, int rightEnd) {
        int leftEnd = (rightEnd + leftStart) / 2;
        int rightStart = leftEnd + 1;
        int size = rightEnd - leftStart + 1;

        int left = leftStart;
        int right = rightStart;
        int index = leftStart;

        while (left <= leftEnd && right <= rightEnd) {
            if (array[left] <= array[right]) {
                temp[index] = array[left];
                left++;
            } else {
                temp[index] = array[right];
                right++;
            }
            index++;
        }

        // Copy the remaining elements
        System.arraycopy(array, left, temp, index, leftEnd - left + 1);
        System.arraycopy(array, right, temp, index, rightEnd - right + 1);

        // Copy sorted elements back to original array
        System.arraycopy(temp, leftStart, array, leftStart, size);
    }
}

